Реформатировал структуру JS файлов Add: Добавлена функция ожидания библиотеки на фронте waitYMaps в index.js Change: Переместил CustomSearchProvider в свой файл Change: Переместил все настройки Яндекс.Карт в файл ymapsSettings.js
